Moved git keymaps
